+++ CHANGES.md
@@ -1,8 +1,9 @@
 **Added:**
-* none
+* clear value support in control properties editor
+* inherit/override buttons in control properties editor
 
 **Changed:**
-* changed methods in ControlClass from returning a list to an Iterable so that list building is faster (using a linked list for them now)
+* none
 
 **Fixed:**
-* control class extending now works completely (no bugs found yet)
+* none